Сделана смена карточки на edit и обратно. Переделан point-presenter. Переделана edit-point-view (собрана в одну view)
